+++ crates/holochain/tests/tests/mod.rs
@@ -5,6 +5,8 @@ mod app_interface_security;
 mod authored_test;
 mod clone_cell;
 // Countersigning HDK functions needed
+#[cfg(feature = "unstable-functions")]
+mod blocks;
 #[cfg(feature = "unstable-countersigning")]
 mod countersigning;
 mod dna_properties;