+++ crates/holochain_p2p/tests/tests/blocks.rs
@@ -1,22 +1,23 @@
 use ::fixt::fixt;
 use holo_hash::{
-    fixt::{AgentPubKeyFixturator, DhtOpHashFixturator, DnaHashFixturator},
+    fixt::{ActionHashFixturator, AgentPubKeyFixturator, DhtOpHashFixturator, DnaHashFixturator},
     DnaHash,
 };
 use holochain_keystore::{test_keystore, MetaLairClient};
 use holochain_p2p::{
     actor::DynHcP2p, event::MockHcP2pHandler, spawn_holochain_p2p, HolochainP2pConfig,
-    HolochainP2pLocalAgent,
+    HolochainP2pError, HolochainP2pLocalAgent,
 };
 use holochain_state::{block::get_all_cell_blocks, prelude::test_conductor_db};
 use holochain_timestamp::{InclusiveTimestampInterval, Timestamp};
 use holochain_types::{
     db::{DbKindConductor, DbKindDht, DbKindPeerMetaStore, DbWrite},
     prelude::{Block, CellBlockReason, CellId},
+    record::WireRecordOps,
 };
 use holochain_zome_types::block::BlockTarget;
 use kitsune2_api::{AgentInfo, AgentInfoSigned, DhtArc, DynBlocks};
-use std::sync::Arc;
+use std::{sync::Arc, time::Duration};
 
 #[tokio::test(flavor = "multi_thread")]
 async fn cell_blocks_are_committed_to_database() {
@@ -86,23 +87,24 @@ async fn block_someone() {
 #[tokio::test(flavor = "multi_thread")]
 async fn agent_is_removed_from_peer_store_when_blocked() {
     let dna_hash = fixt!(DnaHash);
-    let keystore = test_keystore();
-    let agent = keystore.new_sign_keypair_random().await.unwrap();
-    let cell_id = CellId::new(dna_hash.clone(), agent.clone());
 
-    let TestCase { actor, .. } = TestCase::new_with_keystore(&dna_hash, &keystore).await;
-    let space = actor
+    let TestCase { actor: alice, .. } = TestCase::new(&dna_hash).await;
+    let space = alice
         .test_kitsune()
         .space(dna_hash.to_k2_space())
         .await
         .unwrap();
     let peer_store = space.peer_store();
 
-    let local_agent = HolochainP2pLocalAgent::new(agent.clone(), DhtArc::FULL, 1, keystore.clone());
+    // Create Bob's keys so he can be blocked by Alice.
+    let keystore = test_keystore();
+    let bob_pubkey = keystore.new_sign_keypair_random().await.unwrap();
+    let local_agent =
+        HolochainP2pLocalAgent::new(bob_pubkey.clone(), DhtArc::FULL, 1, keystore.clone());
     let agent_info_signed = AgentInfoSigned::sign(
         &local_agent,
         AgentInfo {
-            agent: agent.to_k2_agent(),
+            agent: bob_pubkey.to_k2_agent(),
             created_at: kitsune2_api::Timestamp::now(),
             expires_at: kitsune2_api::Timestamp::from_micros(i64::MAX),
             space: dna_hash.to_k2_space(),
@@ -115,10 +117,15 @@ async fn agent_is_removed_from_peer_store_when_blocked() {
     .unwrap();
     peer_store.insert(vec![agent_info_signed]).await.unwrap();
 
-    assert!(peer_store.get(agent.to_k2_agent()).await.unwrap().is_some());
+    assert!(peer_store
+        .get(bob_pubkey.to_k2_agent())
+        .await
+        .unwrap()
+        .is_some());
 
-    // Block an agent.
-    actor
+    // Alice blocks Bob.
+    let cell_id = CellId::new(dna_hash.clone(), bob_pubkey.clone());
+    alice
         .block(Block::new(
             BlockTarget::Cell(cell_id, CellBlockReason::BadCrypto),
             InclusiveTimestampInterval::try_new(Timestamp::now(), Timestamp::max()).unwrap(),
@@ -126,8 +133,12 @@ async fn agent_is_removed_from_peer_store_when_blocked() {
         .await
         .unwrap();
 
-    // Check agent has been removed from peer store.
-    assert!(peer_store.get(agent.to_k2_agent()).await.unwrap().is_none());
+    // Check Bob has been removed from Alice's peer store.
+    assert!(peer_store
+        .get(bob_pubkey.to_k2_agent())
+        .await
+        .unwrap()
+        .is_none());
 }
 
 #[tokio::test(flavor = "multi_thread")]
@@ -325,6 +336,95 @@ async fn block_is_scoped_per_dna() {
         .unwrap());
 }
 
+#[tokio::test(flavor = "multi_thread")]
+async fn get_from_blocked_agent_fails() {
+    let dna_hash = DnaHash::from_raw_32(vec![0xaa; 32]);
+    let keystore_1 = test_keystore();
+    let keystore_2 = test_keystore();
+    let TestCase { actor: alice, .. } = TestCase::new_with_keystore(&dna_hash, &keystore_1).await;
+    let TestCase { actor: bob, .. } = TestCase::new_with_keystore(&dna_hash, &keystore_2).await;
+    let alice_pubkey = keystore_1.new_sign_keypair_random().await.unwrap();
+    let bob_pubkey = keystore_2.new_sign_keypair_random().await.unwrap();
+    alice
+        .join(dna_hash.clone(), alice_pubkey.clone(), None)
+        .await
+        .unwrap();
+    bob.join(dna_hash.clone(), bob_pubkey.clone(), None)
+        .await
+        .unwrap();
+    bob.test_set_full_arcs(dna_hash.to_k2_space()).await;
+
+    // Exchange peer infos to accelerate bootstrapping.
+    tokio::time::timeout(Duration::from_secs(10), async {
+        loop {
+            let maybe_agent_info_1 = alice
+                .peer_store(dna_hash.clone())
+                .await
+                .unwrap()
+                .get(alice_pubkey.to_k2_agent())
+                .await
+                .unwrap();
+            let maybe_agent_info_2 = bob
+                .peer_store(dna_hash.clone())
+                .await
+                .unwrap()
+                .get(bob_pubkey.to_k2_agent())
+                .await
+                .unwrap();
+            if let (Some(agent_info_1), Some(agent_info_2)) =
+                (maybe_agent_info_1, maybe_agent_info_2)
+            {
+                alice
+                    .peer_store(dna_hash.clone())
+                    .await
+                    .unwrap()
+                    .insert(vec![agent_info_2])
+                    .await
+                    .unwrap();
+                bob.peer_store(dna_hash.clone())
+                    .await
+                    .unwrap()
+                    .insert(vec![agent_info_1])
+                    .await
+                    .unwrap();
+                break;
+            }
+            tokio::time::sleep(Duration::from_millis(100)).await;
+        }
+    })
+    .await
+    .unwrap();
+
+    let response = alice.get(dna_hash.clone(), fixt!(ActionHash).into()).await;
+    assert!(response.is_ok());
+
+    alice
+        .block(Block::new(
+            BlockTarget::Cell(
+                CellId::new(dna_hash.clone(), bob_pubkey.clone()),
+                CellBlockReason::BadCrypto,
+            ),
+            InclusiveTimestampInterval::try_new(
+                holochain_timestamp::Timestamp::now(),
+                holochain_timestamp::Timestamp::max(),
+            )
+            .unwrap(),
+        ))
+        .await
+        .unwrap();
+
+    // Blocking removes an agent from the peer store.
+    // To make sure the block is being enforced, the agent should be re-added to the
+    // peer 1's peer store. That isn't possible, however, because the peer store
+    // implementation internally discards blocked agents when inserting.
+
+    let response = alice.get(dna_hash.clone(), fixt!(ActionHash).into()).await;
+    assert!(matches!(
+        response,
+        Err(HolochainP2pError::NoPeersForLocation(_, _))
+    ));
+}
+
 struct TestCase {
     actor: DynHcP2p,
     blocks_module: DynBlocks,
@@ -373,10 +473,15 @@ impl TestCase {
             ..Default::default()
         };
         let actor = spawn_holochain_p2p(config, keystore).await.unwrap();
-        actor
-            .register_handler(Arc::new(MockHcP2pHandler::new()))
-            .await
-            .unwrap();
+        let mut handler = MockHcP2pHandler::new();
+        handler.expect_handle_get().returning(|_, _, _| {
+            Box::pin(async move {
+                Ok(holochain_types::dht_op::WireOps::Record(
+                    WireRecordOps::new(),
+                ))
+            })
+        });
+        actor.register_handler(Arc::new(handler)).await.unwrap();
         let space = actor
             .test_kitsune()
             .space(dna_hash.to_k2_space())